Fix move env logic in index.js
review suggestion by @RishuKumarCodes
