--- Server/Config/envConfig.js
@@ -1,5 +1,0 @@
-import dotenv from "dotenv";
-dotenv.config();
-export const PORT = Number(process.env.PORT) || 3000;
-export const NODE_ENV = process.env.NODE_ENV || "development";
-export const CORS_ORIGIN = process.env.CORS_ORIGIN || "http://localhost:5173";
